refactor(typescript): move cli-builder into shared source
