+++ code/packages/typescript/cli-builder/vitest.config.ts
@@ -3,6 +3,12 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     coverage: {
+      include: [
+        "src/**/*.ts",
+      ],
+      exclude: [
+        "src/index.ts",
+      ],
       provider: "v8",
       thresholds: {
         lines: 90,
